Renamed tests to conform to convention
