+++ core/src/test/java/org/mule/runtime/core/internal/processor/strategy/StreamPerEventSinkTestCase.java
@@ -20,13 +20,12 @@ import java.util.function.Consumer;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class StreamPerEventSinkTest {
+class StreamPerEventSinkTestCase {
 
   private StreamPerEventSink sink;
   @Mock
