+++ core/src/test/java/org/mule/runtime/core/internal/processor/strategy/reactor/builder/PipelineProcessingStrategyReactiveProcessorBuilderUnitTestCase.java
@@ -11,44 +11,37 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.mule.runtime.api.scheduler.Scheduler;
-import org.mule.runtime.core.api.event.CoreEvent;
 import org.mule.runtime.core.api.processor.ReactiveProcessor;
-import org.reactivestreams.Publisher;
 import reactor.test.publisher.TestPublisher;
 
 import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Parameter;
 import java.util.Arrays;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class ComponentProcessingStrategyReactiveProcessorBuilderTest {
+class PipelineProcessingStrategyReactiveProcessorBuilderUnitTestCase {
 
-  private ComponentProcessingStrategyReactiveProcessorBuilder builder;
+  private PipelineProcessingStrategyReactiveProcessorBuilder builder;
+
   @Mock
-  private Scheduler scheduler;
-  @Mock
-  private ReactiveProcessor processor;
-  @Mock
-  private Publisher<CoreEvent> publisher;
+  private ReactiveProcessor pipeline;
 
   @BeforeEach
   void setUp() {
-    builder = ComponentProcessingStrategyReactiveProcessorBuilder.processingStrategyReactiveProcessorFrom(processor, scheduler,
-                                                                                                          "appId", "app");
+    when(pipeline.apply(any())).thenAnswer(inv -> inv.getArgument(0));
+
+    builder = PipelineProcessingStrategyReactiveProcessorBuilder
+        .pipelineProcessingStrategyReactiveProcessorFrom(pipeline, ClassLoader.getPlatformClassLoader(), "id", "app");
   }
 
   @Test
-  void build() {
-    when(processor.apply(any())).thenAnswer(inv -> inv.getArgument(0));
-    Arrays.stream(ComponentProcessingStrategyReactiveProcessorBuilder.class.getDeclaredMethods())
+  void pipelineProcessingStrategyReactiveProcessorFrom() {
+    Arrays.stream(PipelineProcessingStrategyReactiveProcessorBuilder.class.getDeclaredMethods())
         .filter(m -> m.getName().startsWith("with"))
         .filter(m -> m.canAccess(builder))
         .forEach(m -> {
